Made method, algorithm() use a StringBuilder rather than a String
The first idea that came to mind when trying to optimize. Very minor change and I can't tell if there was much of a change. The timer is ± 1 ms having times range from around 2.5 ms to 3.5 from what I have seen from running it on a file many times. I will try other ideas to optimize.

After more time spent with the code and understanding it, there are more tasks to be done. Namely:

- editing the decompression method to read in compressed file outputs and use data from the files rather than stored private instance variables.